Extract the backports logic into its own helper class
