+++ bots/notify/src/test/java/org/openjdk/skara/bots/notify/BackportsTests.java
@@ -1,0 +1,100 @@
+/*
+ * Copyright (c) 2020, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+package org.openjdk.skara.bots.notify;
+
+import org.junit.jupiter.api.*;
+import org.openjdk.skara.issuetracker.Link;
+import org.openjdk.skara.json.JSON;
+import org.openjdk.skara.test.HostCredentials;
+
+import java.io.IOException;
+import java.util.Optional;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class BackportsTests {
+    @Test
+    void mainIssue(TestInfo testInfo) throws IOException {
+        try (var credentials = new HostCredentials(testInfo)) {
+            var issueProject = credentials.getIssueProject();
+
+            var issue1 = credentials.createIssue(issueProject, "Issue 1");
+            issue1.setProperty("issuetype", JSON.of("Bug"));
+
+            var issue2 = credentials.createIssue(issueProject, "Issue 2");
+            issue2.setProperty("issuetype", JSON.of("Backport"));
+            issue1.addLink(Link.create(issue2, "backported by").build());
+
+            var issue3 = credentials.createIssue(issueProject, "Issue 3");
+            issue3.setProperty("issuetype", JSON.of("Backport"));
+            issue3.addLink(Link.create(issue1, "backport of").build());
+
+            assertEquals(issue1, Backports.findMainIssue(issue1).orElseThrow());
+            assertEquals(issue1, Backports.findMainIssue(issue2).orElseThrow());
+            assertEquals(issue1, Backports.findMainIssue(issue3).orElseThrow());
+        }
+    }
+
+    @Test
+    void noMainIssue(TestInfo testInfo) throws IOException {
+        try (var credentials = new HostCredentials(testInfo)) {
+            var issueProject = credentials.getIssueProject();
+
+            var issue1 = credentials.createIssue(issueProject, "Issue 1");
+            issue1.setProperty("issuetype", JSON.of("Bug"));
+
+            var issue2 = credentials.createIssue(issueProject, "Issue 2");
+            issue2.setProperty("issuetype", JSON.of("Backport"));
+
+            var issue3 = credentials.createIssue(issueProject, "Issue 3");
+            issue3.setProperty("issuetype", JSON.of("Backport"));
+            issue2.addLink(Link.create(issue3, "backported by").build());
+
+            assertEquals(issue1, Backports.findMainIssue(issue1).orElseThrow());
+            assertEquals(Optional.empty(), Backports.findMainIssue(issue2));
+            assertEquals(Optional.empty(), Backports.findMainIssue(issue3));
+        }
+    }
+
+    @Test
+    void nonBackportLink(TestInfo testInfo) throws IOException {
+        try (var credentials = new HostCredentials(testInfo)) {
+            var issueProject = credentials.getIssueProject();
+
+            var issue1 = credentials.createIssue(issueProject, "Issue 1");
+            issue1.setProperty("issuetype", JSON.of("Bug"));
+
+            var issue2 = credentials.createIssue(issueProject, "Issue 2");
+            issue2.setProperty("issuetype", JSON.of("Bug"));
+            issue1.addLink(Link.create(issue2, "duplicated by").build());
+
+            var issue3 = credentials.createIssue(issueProject, "Issue 3");
+            issue3.setProperty("issuetype", JSON.of("CSR"));
+            issue1.addLink(Link.create(issue3, "CSRed by").build());
+
+            assertEquals(issue1, Backports.findMainIssue(issue1).orElseThrow());
+            assertEquals(issue2, Backports.findMainIssue(issue2).orElseThrow());
+            assertEquals(Optional.empty(), Backports.findMainIssue(issue3));
+        }
+    }
+}
